requirements.txt
Flask==2.3.3
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.1
flask-cors==4.0.0
werkzeug==2.3.7
